Document codegen crate and break out transform logic into its own file
